Add S3 storage class configuration option
A new environment variable called `WALG_S3_STORAGE_CLASS` allows
users to set the storage class used when wal-g writes objects to
S3. This allows usage of the Infrequent Access or Reduced
Redundancy features of S3.

In keeping with current behaviour, new objects are created with the
"STANDARD" storage class if the variable is not set.