Backups on replication slaves (#32)
Allow backup from replication slave. In this case Timeline is read from pg_control_snapshot() and formatted in Go. This functionality will not work unless you have default physical parameters of WAL (specifically, segment size). The compatibility is checked by pg_control_init() call.
If the server is promoted during backup or receives timeline switch command, backup will be uploaded, but sentinel.json will not. User can create empty json file and enable backup-fetch for this copy. This functionality is kept in fear that timeline switch and server promotion is often an indication that hardware is failing and fresh copy can be handy, even if it's use require a lot more of precaution work.