Avoid assumption of pg_control sentinel
WAL-G uploads pg_control separately so that it can be extracted at the
end of a successful backup restoration.  Postgres cannot start up with
pg_control, and this is to prevent partial downloads from causing
damage because the database may appear to start normally.

However, WAL-E doesn't use this trick, and the slick use of lexical
order of the keys as returned from S3 would truncate off one
archive(!).  So use a regular expression to match the pg_control
archive and make it optional.