Fix prevent wal overwrite (#410)
* Add test for PostgreSQL WALG_PREVENT_WAL_OVERWRITE

This is currently broken so a test for it that fails with the current
implementation is needed. We can then fix the bahviour and the test
should pass.

* Fix WALG_PREVENT_WAL_OVERWRITE

The compressors file extention was being appended twice preventing us
from correctly detecting overwrites.

* Don't push PostgreSQL WAL if already archived

If we detect that a PostgreSQL WAL file has already been archived, and
it has equal content, then we don't need to upload it again.