Fix release tarball creation to use zero uid and gid (#1025)
Fixes the release action so that the tarball creation uses a zero uid and gid
value for files inside the release archive. Otherwise the archive will inherit
the uid/gid of the GitHub Action runner and those values will show up on end
user systems if the archive is extracted as root.