sql server - skipping logs that already been applied
Image the case - db is backup is restored, some logs applied and then something bad happens and you're stuck. To proceed with logs restore you basically have to drop database and start over. If you try to run logs-restore again you'll get and error, as wal-g will try to apply all the logs since last full backup.
Additional check can mitigate this. Check the LSNs of the log backup and the database in  the 'restoring' state and skipping those whose LSN is somewhere in the past for the current sate of the recovery.