PostgreSQL - Reduce memory usage by not tracking metadata of the files backed up (#1101)
* PostgreSQL - Reduce memory usage by not tracking metadata of the files backed up

By default, WAL-G tracks metadata of the files backed up. If millions of files are backed up (typically in case of hundreds of databases and thousands of tables in each database), tracking this metadata alone would require GBs of memory.

If `--reduce-memory-usage` or `WALG_REDUCE_MEMORY_USAGE` is enabled, WAL-G does not track metadata of the files backed up. This significantly reduces the memory usage on instances with `> 100k` files.

Limitations

* Cannot be used with `rating-composer`, `copy-composer`
* Cannot be used with `delta-from-user-data`, `delta-from-name`, `add-user-data`

* Rename --reduce-memory-usage (WALG_REDUCE_MEMORY_USAGE) flag to --without-files-metadata (WALG_WITHOUT_FILES_METADATA)

* Minor fixes

Co-authored-by: usernamedt <usernamedt@protonmail.ch>