Support profiling by not stripping binaries
When building the wal-g binary, previously we were telling
the linker to strip the symbol table and debug sections
from the binary.  Doing this badly inhibits observability.
It breaks profiling and dynamic instrumentation, making it
more difficult to analyze and improve performance bottlenecks.

Stripping the binary saves roughly 14 MB.  In general that seems
like a small cost in exchange for greatly improved observability.

If anyone using wal-g would prefer to have that space savings
rather than the ability to profile, they can reapply the stripping
either:
* during the build, by setting the make variable STRIP_BINARIES
* after the build, by using "strip" or "objcopy --strip-debug"