Fix bug in backup-fetch (#1537)
I forgot to update the folder object after setting new policies.
Fortunately, the default policies were almost the same, so it led to only a few problems. For example, if two backups existed in different storages with the same name and LATEST backup was fetched, the backup from the first storage was returned even if the backup from the second was newer.