Fix bug in memory folder
If we create a folder with path "/" (representing the root folder), or with any other path starting with "/", the listing breaks: it finds no files or subfolders because their paths don't start with "/".

Also stop creating multi-storage folders with path "/", and always represent the root path by "" to avoid ambiguity.