Fixed dropIndexes operation replay.
To replay a dropIndexes operation using the applyOps command, MongoDB needs the collection UUID (the "ui" field in an oplog document).
However, WAL-G erases all UUIDs before pushing an oplog.
This is a correct strategy (since operations are meant to be replayable without UUIDs), but it clashes with the current Mongo behavior when dropping indexes.

In Mongo versions prior to 7.0, such a command (without UUID) has no effect: it completes successfully, but no indexes are dropped.
In Mongo 7.0.1-7.0.2, there is a bug (https://jira.mongodb.org/browse/SERVER-82349) that leads to server crash on such commands.

So, this commit solves two issues:

* it fixes dropIndexes operation replay on all the Mongo versions (by handling them separately, without using applyOps)

* it helps support Mongo 7.0, since without this fix an oplog replay may crash the server thread and drop the connection

Before this commit, index drop and corresponding collection drop were performed in a single scenario step, so we did not catch this bug earlier.
In this commit, new steps are added to verify that an index is dropped correctly.

Since index drop works now, we can face a BackgroundOperationInProgressForNamespace on Mongo versions prior to 5.2.
Ignored this error and documented this behavior.