Fix delta file not exist in getDeltaMap
If the first wal file pushed by the wal-g is not started as xxx0
after the archive_mode enabled, and we start the basebackup, then
the first delta segment will be missing for that basebackup. So
does the last delta file, if the last delta segment is the same
as the first delta segment.

The first wal log file after starting the basebackup doesn't
contain XLogRecord tail in the head, so we can parse the wal files
to extract locations.

Add test 'delta_backup_wal_delta_missing_test.sh' for this commit