Fix getWalSegmentRange
In 'getWalSegmentRange', the redo point points to a position in the wal
file. The delta backup will wait for that wal log file, but the wal log
file which contains checkpoint and backup-end XlogRecord will be archived
only after the backup finished, so the delta backup will fail. This is a
common case when this is a few insert traffics in the database.

However there is no good method to solve it now. Fallback to the fullbackup
may be the best choice.

For function 'getWalSegmentRange', I think should change
'lastUsedLsn := firstNotUsedLsn - 1'
to
'lastUsedLsn := firstNotUsedLsn - SizeOfXLogLongPHD - 1'
as the redo point is after the XLogLongPageHeaderData

This can make delta backup work if the redo point starts at the begin of
a wal file.